Add i18n for some map controls and improve dummy locales structure
