+++ addon/locales/en/components/flexberry-map.js
@@ -1,3 +1,11 @@
 export default {
-  'queryText': 'Search objects'
+  'queryText': 'Search objects',
+  'zoom-control': {
+    'zoom-in-button': {
+      'title': 'Zoom in'
+    },
+    'zoom-out-button': {
+      'title': 'Zoom out'
+    }
+  }
 };
