+++ addon/locales/en/components/flexberry-maplayer.js
@@ -1,3 +1,23 @@
 export default {
-  'opacity': 'Visibility'
+  'opacity-label': {
+    'tooltip': 'Layer visibility'
+  },
+  'attributes-button': {
+    'tooltip': 'Show layer attributes panel'
+  },
+  'bounds-button': {
+    'tooltip': 'Fit layer bounds'
+  },
+  'add-button': {
+    'tooltip': 'Add new child layer'
+  },
+  'copy-button': {
+    'tooltip': 'Create layer copy'
+  },
+  'edit-button': {
+    'tooltip': 'Edit layer settings'
+  },
+  'remove-button': {
+    'tooltip': 'Remove layer'
+  }
 };
