+++ addon/locales/en/components/history-control.js
@@ -1,0 +1,8 @@
+export default {
+  'back-button': {
+    'title': 'Previous extent'
+  },
+  'forward-button': {
+    'title': 'Next extent'
+  }
+};
