+++ addon/locales/ru/components/flexberry-map.js
@@ -1,3 +1,11 @@
 export default {
-  'queryText': 'Поиск объектов'
+  'queryText': 'Поиск объектов',
+  'zoom-control': {
+    'zoom-in-button': {
+      'title': 'Приблизить'
+    },
+    'zoom-out-button': {
+      'title': 'Отдалить'
+    }
+  }
 };
