+++ addon/locales/ru/components/flexberry-maplayer.js
@@ -1,3 +1,23 @@
 export default {
-  'opacity': 'Видимость'
+  'opacity-label': {
+    'tooltip': 'Видимость слоя'
+  },
+  'attributes-button': {
+    'tooltip': 'Показать панель атрибутов слоя'
+  },
+  'bounds-button': {
+    'tooltip': 'Приблизить к границам слоя'
+  },
+  'add-button': {
+    'tooltip': 'Добавить новй дочерний слой'
+  },
+  'copy-button': {
+    'tooltip': 'Создать копию слоя'
+  },
+  'edit-button': {
+    'tooltip': 'Редактировать настройки слоя'
+  },
+  'remove-button': {
+    'tooltip': 'Удалить слой'
+  }
 };
