+++ addon/locales/ru/components/history-control.js
@@ -1,0 +1,8 @@
+export default {
+  'back-button': {
+    'title': 'Предыдущий экстент'
+  },
+  'forward-button': {
+    'title': 'Следующий экстент'
+  }
+};
