+++ addon/templates/components/flexberry-maplayer.hbs
@@ -30,22 +30,26 @@
           {{flexberry-button
             class=(concat flexberryClassNames.removeButton " right floated " flexberryClassNames.preventExpandCollapse)
             iconClass="trash icon"
+            tooltip=(t "components.flexberry-maplayer.remove-button.tooltip")
             click=(action "onRemoveButtonClick")
           }}
           {{flexberry-button
             class=(concat flexberryClassNames.editButton " right floated " flexberryClassNames.preventExpandCollapse)
             iconClass="edit icon"
+            tooltip=(t "components.flexberry-maplayer.edit-button.tooltip")
             click=(action "onEditButtonClick")
           }}
           {{flexberry-button
             class=(concat flexberryClassNames.copyButton " right floated " flexberryClassNames.preventExpandCollapse)
             iconClass="copy icon"
+            tooltip=(t "components.flexberry-maplayer.copy-button.tooltip")
             click=(action "onCopyButtonClick")
           }}
           {{#if _addOperationIsAvailable}}
             {{flexberry-button
               class=(concat flexberryClassNames.addButton " right floated " flexberryClassNames.preventExpandCollapse)
               iconClass="plus icon"
+              tooltip=(t "components.flexberry-maplayer.add-button.tooltip")
               click=(action "onAddButtonClick")
             }}
           {{/if}}
@@ -53,6 +57,7 @@
             {{flexberry-button
               class=(concat flexberryClassNames.boundsButton " right floated " flexberryClassNames.preventExpandCollapse)
               iconClass="maximize icon"
+              tooltip=(t "components.flexberry-maplayer.bounds-button.tooltip")
               click=(action "onBoundsButtonClick")
             }}
           {{/if}}
@@ -60,12 +65,13 @@
             {{flexberry-button
               class=(concat flexberryClassNames.attributesButton " right floated " flexberryClassNames.preventExpandCollapse)
               iconClass="browser icon"
+              tooltip=(t "components.flexberry-maplayer.attributes-button.tooltip")
               click=(action "onAttributesButtonClick")
             }}
           {{/if}}
         {{/if}}
         {{#if _canChangeOpacity}}
-          <label class="{{flexberryClassNames.opacityLabel}}" title="{{t "components.flexberry-maplayer.opacity" }}">
+          <label class="{{flexberryClassNames.opacityLabel}}" title="{{t "components.flexberry-maplayer.opacity-label.tooltip"}}">
             {{_opacityDisplayValue}}
           </label>
         {{/if}}
