+++ tests/dummy/app/locales/en/forms/application/footer.js
@@ -1,0 +1,9 @@
+export default {
+  'application-name': 'Test stand for ember-flexberry-gis',
+  'application-version': {
+    'caption': 'Addon version {{version}}',
+    'title': 'It is version of ember-flexberry-gis addon, which uses in this dummy application ' +
+      '(npm version + commit sha). ' +
+      'Click to open commit on GitHub.'
+  }
+};
