+++ tests/dummy/app/locales/en/forms/application/header.js
@@ -1,0 +1,21 @@
+export default {
+  'menu': {
+    'sitemap-button': {
+      'caption': '',
+      'title': 'Menu'
+    },
+    'user-settings-service-checkbox': {
+      'caption': 'Use user settings service'
+    },
+    'show-menu': {
+      'caption': 'Show menu'
+    },
+    'hide-menu': {
+      'caption': 'Hide menu'
+    },
+    'language-dropdown': {
+      'caption': 'Application language',
+      'placeholder': 'Choose language'
+    }
+  }
+};
