+++ tests/dummy/app/locales/en/forms/application/sitemap.js
@@ -1,0 +1,48 @@
+export default {
+  'application-name': {
+    'caption': 'Test stand for ember-flexberry-gis',
+    'title': ''
+  },
+  'application-version': {
+    'caption': 'Addon version {{version}}',
+    'title': 'It is version of ember-flexberry-gis addon, which uses in this dummy application ' +
+      '(npm version + commit sha). ' +
+      'Click to open commit on GitHub.'
+  },
+  'index': {
+    'caption': 'Home',
+    'title': ''
+  },
+  'gis': {
+    'caption': 'GIS',
+    'title': '',
+    'maps': {
+      'caption': 'Maps',
+      'title': ''
+    },
+    'csw-connections': {
+      'caption': 'CSW connections',
+      'title': ''
+    },
+    'map-metadata': {
+      'caption': 'Мetadata',
+      'title': ''
+    },
+    'gis-search-form': {
+      'caption': 'Search',
+      'title': 'Search through layer metadata and maps'
+    }
+  },
+  'components-examples': {
+    'caption': 'Components examples',
+    'title': '',
+    'flexberry-boundingbox': {
+      'caption': 'flexberry-boundingbox',
+      'title': '',
+      'settings-example': {
+        'caption': 'Пример работы компонента',
+        'title': ''
+      }
+    }
+  }
+};
