+++ tests/dummy/app/locales/en/forms/components-examples.js
@@ -1,0 +1,7 @@
+export default {
+  'flexberry-boundingbox': {
+    'settings-example': {
+      'caption': 'Flexberry-boundingbox testing form'
+    }
+  }
+};
