+++ tests/dummy/app/locales/en/forms/edit-form.js
@@ -1,0 +1,8 @@
+export default {
+  'save-success-message-caption': 'Save operation succeed',
+  'save-success-message': 'Object saved',
+  'save-error-message-caption': 'Save operation failed',
+  'delete-success-message-caption': 'Delete operation succeed',
+  'delete-success-message': 'Object deleted',
+  'delete-error-message-caption': 'Delete operation failed'
+};
