+++ tests/dummy/app/locales/en/forms/index.js
@@ -1,0 +1,3 @@
+export default {
+  'greeting': 'Welcome to ember-flexberry-gis test stand!'
+};
