+++ tests/dummy/app/locales/en/forms/loading.js
@@ -1,0 +1,3 @@
+export default {
+  'spinner-caption': 'Loading data, please wait...'
+};
