+++ tests/dummy/app/locales/en/forms/maps.js
@@ -1,0 +1,3 @@
+export default {
+  'caption': 'Maps'
+};
