+++ tests/dummy/app/locales/ru/forms.js
@@ -1,0 +1,17 @@
+import Index from './forms/index';
+import Loading from './forms/loading';
+import Application from './forms/application';
+import EditForm from './forms/edit-form';
+import Maps from './forms/maps';
+import Map from './forms/map';
+import ComponentsExamples from './forms/components-examples';
+
+export default {
+  'index': Index,
+  'loading': Loading,
+  'application': Application,
+  'edit-form': EditForm,
+  'maps': Maps,
+  'map': Map,
+  'components-examples': ComponentsExamples
+};
