+++ tests/dummy/app/locales/ru/forms/application.js
@@ -1,0 +1,9 @@
+import Header from './application/header';
+import Footer from './application/footer';
+import Sitemap from './application/sitemap';
+
+export default {
+  'header': Header,
+  'footer': Footer,
+  'sitemap': Sitemap
+};
