+++ tests/dummy/app/locales/ru/forms/application/footer.js
@@ -1,0 +1,9 @@
+export default {
+  'application-name': 'Тестовый стенд ember-flexberry-gis',
+  'application-version': {
+    'caption': 'Версия аддона {{version}}',
+    'title': 'Это версия аддона ember-flexberry-gis, которая сейчас используется в этом тестовом приложении ' +
+      '(версия npm-пакета + хэш коммита). ' +
+      'Кликните, чтобы перейти на GitHub.'
+  }
+};
