+++ tests/dummy/app/locales/ru/forms/application/header.js
@@ -1,0 +1,21 @@
+export default {
+  'menu': {
+    'sitemap-button': {
+      'caption': '',
+      'title': 'Меню'
+    },
+    'user-settings-service-checkbox': {
+      'caption': 'Использовать сервис пользовательских настроек'
+    },
+    'show-menu': {
+      'caption': 'Показать меню'
+    },
+    'hide-menu': {
+      'caption': 'Скрыть меню'
+    },
+    'language-dropdown': {
+      'caption': 'Язык приложения',
+      'placeholder': 'Выберете язык'
+    }
+  }
+};
