+++ tests/dummy/app/locales/ru/forms/application/sitemap.js
@@ -1,0 +1,48 @@
+export default {
+  'application-name': {
+    'caption': 'Тестовый стенд ember-flexberry-gis',
+    'title': ''
+  },
+  'application-version': {
+    'caption': 'Версия аддона {{version}}',
+    'title': 'Это версия аддона ember-flexberry-gis, которая сейчас используется в этом тестовом приложении ' +
+      '(версия npm-пакета + хэш коммита). ' +
+      'Кликните, чтобы перейти на GitHub.'
+  },
+  'index': {
+    'caption': 'Главная',
+    'title': ''
+  },
+  'gis': {
+    'caption': 'ГИС',
+    'title': '',
+    'maps': {
+      'caption': 'Карты',
+      'title': ''
+    },
+    'csw-connections': {
+      'caption': 'CSW соединения',
+      'title': ''
+    },
+    'map-metadata': {
+      'caption': 'Метаданные',
+      'title': ''
+    },
+    'gis-search-form': {
+      'caption': 'Поиск',
+      'title': 'Поиск по метаданным слоёв и проектам карт'
+    }
+  },
+  'components-examples': {
+    'caption': 'Примеры компонентов',
+    'title': '',
+    'flexberry-boundingbox': {
+      'caption': 'flexberry-boundingbox',
+      'title': '',
+      'settings-example': {
+        'caption': 'Пример работы компонента',
+        'title': ''
+      }
+    }
+  }
+};
