+++ tests/dummy/app/locales/ru/forms/components-examples.js
@@ -1,0 +1,7 @@
+export default {
+  'flexberry-boundingbox': {
+    'settings-example': {
+      'caption': 'Тестовая форма flexberry-boundingbox'
+    }
+  }
+};
