+++ tests/dummy/app/locales/ru/forms/edit-form.js
@@ -1,0 +1,8 @@
+export default {
+  'save-success-message-caption': 'Сохранение завершилось успешно',
+  'save-success-message': 'Объект сохранен',
+  'save-error-message-caption': 'Ошибка сохранения',
+  'delete-success-message-caption': 'Удаление завершилось успешно',
+  'delete-success-message': 'Объект удален',
+  'delete-error-message-caption': 'Ошибка удаления'
+};
