+++ tests/dummy/app/locales/ru/forms/index.js
@@ -1,0 +1,3 @@
+export default {
+  'greeting': 'Добро пожаловать на тестовый стенд ember-flexberry-gis!'
+};
