+++ tests/dummy/app/locales/ru/forms/loading.js
@@ -1,0 +1,3 @@
+export default {
+  'spinner-caption': 'Данные загружаются, пожалуйста подождите...'
+};
