+++ tests/dummy/app/locales/ru/forms/maps.js
@@ -1,0 +1,3 @@
+export default {
+  'caption': 'Карты'
+};
