+++ tests/dummy/app/templates/map.hbs
@@ -176,7 +176,6 @@
           leafletMap=leafletMap
         }}
         {{map-commands/go-to
-          availableCRS=availableCRS
           execute=(action "onMapCommandExecute" target=mapToolbar)
         }}
         {{map-commands/locate
